Move probabilisticsamplerprocessor one level up
