+++ processor/probabilisticsamplerprocessor/factory.go
