+++ processor/probabilisticsamplerprocessor/testdata/empty.yaml
